move umami class to model directory
